refactor: disable code minification and enable resource shrinking in release build
The changes disable code minification (R8) by setting `minifyEnabled` to `false` in the release build type within `build.gradle`, and they enable resource shrinking by setting `shrinkResources` to `true`. In `proguard-rules.pro` file the comment was modified to be more specific, now it indicates that the added classes are based on the error log.